backend/requirements.txt
Django==5.1.5
gunicorn==21.2.0
whitenoise==6.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-decouple==3.8
pandas==2.2.3
numpy==2.2.6
scikit-learn==1.6.1
